+++ src/fct/slt.c
@@ -13,7 +13,8 @@
 int fct_slt(inst k, mem memory, registre* regs)
 {
 	 
-	if (regs[k.r.rs].val<regs[k.r.rt].val) {regs[k.i.rd].val=1;}
+	if ((signed int)regs[k.r.rs].val<(signed int)regs[k.r.rt].val) 
+		{regs[k.i.rd].val=1;}
 	else {regs[k.i.rd].val=0;}
 	return cmd_ok;
 }
