+++ src/fct/sltu.c
@@ -13,7 +13,7 @@
 int fct_sltu(inst k, mem memory, registre* regs)
 {
 	 
-	if (regs[k.r.rs].val<regs[k.r.rt].val) {regs[k.r.rd].val=1;}
+	if ((unsigned int)regs[k.r.rs].val<(unsigned int)regs[k.r.rt].val) {regs[k.r.rd].val=1;}
 	else {regs[k.r.rd].val=0;}
 	return cmd_ok;
 }
